+++ client/src/components/ui/index.ts
@@ -1,6 +1,6 @@
 import { Avatar, AvatarFallback, AvatarImage } from "./avatar";
 import { Button } from "./button";
-import Search from "./search";
+import Search from "../search/search";
 
 export {
     Avatar,
